Added "Node ID" to vtk write function
Added functionality to write a unique Node ID to the VTK output data.
This is useful for doing temporal particle tracking in Paraview. Also
added the option to read in "Node ID" data from a CACTUS output file.

Also changed the name of the "Element" column in the Wake Element data
to "Origin Node", which reflects the most current version of CACTUS
output.