Changed to read in time-separated data
Changed pyCactus.py and pyCactusWake.py to read in time-separated data.
This is a major change to the post-processing workflow, which previously
expected all the time data to be in a single CSV file. This had memory
limitations for large data files.

In the new workflow, the CACTUS CSV data for wake grid and wake nodes is
expected to be split into files by time. The CactusWake and
CactusWakeGrid classes will read in single files at a time. These files
can be batch converted to VTK time-series using the
write_to_vtk_series() class functions. Instantaneous data can be
returned using the get_df_inst() and wakedata_from_df() class functions.

Information about the time and filenames is stored in a class-variable
dictionary.

For legacy monolithic CSV files that contain all the timesteps, the
split_file_by_time.py script can be used.