Simplified loading of data on instantiation
Fixed a problem where, to save memory, rev, element, param, and time
data was only loaded when called. This required each data to be called
explicitly before it would be loaded. However the memory savings
weren't worth the inconvenience. Now, all the rev, element, param, and
time data is loaded upon instantiation of the CactusRun class.