Update pyCactusCsvToVtk.py
Update script which converts CSV wake element and field data to VTK.
Added optional parameters to allow the user to enter glob search strings
from the command line. Also added an option to allow the user to specify
the path to the pyCactus module location. Renamed the calls to CactusRun
methods and attributes to match the current names.