Fix bug in CactusField init
Fix a bug in CactusField init which was introduced in 7fff4ce which
was causing multiple filenames to be passed to read_grid_dims(), when it
should only take one. The fix puts the statement in a for loop which
breaks after first execution. This approach is safe even when
self.filenames is an empty list.